+++ datalab/gui/panel/image.py
@@ -117,7 +117,8 @@ class ImagePanel(BaseDataPanel[ImageObj, ImageROI, roieditor.ImageROIEditor]):
             newparam = NewImageParam()
         if title is not None:
             newparam.title = title
-        if curobj is not None:
+        if curobj is not None and Conf.proc.use_image_dims.get(True):
+            # Use current image dimensions for new image:
             newparam.height, newparam.width = curobj.data.shape
             newparam.dtype = ImageDatatypes.from_dtype(curobj.data.dtype)
         return newparam